+++ src/main/java/com/example/Tollpay/repository/TravellerRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 //This CrudRepository Class Provided by JPA which provide us all function to perform on databases
 //with our entity data
 //for more clarity check class defination
-public interface TravellerRepository extends JpaRepository<Traveller,Long> {
+public interface TravellerRepository extends JpaRepository<Traveller,Integer> {
     public Traveller findByEmail(String email);
     @Query(value = "update TRAVELLER  set WALLET_AMOUNT=?1 where EMAIL= ?2")
     public void updateByEmail(Double amount,String email);
